ngrx store with snackbar
I have implemented ngrx store in this demo project as an assignment, it is working fine as described in the documentation.
from the header section you can collapse/ expand the side bar, with the delete button you can delete the selected items, with add new item you can add new item on the top of the list. This assignment is done for desktop view only.